% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-integrate.R
\name{integrate_discont_val}
\alias{integrate_discont_val}
\title{Integrate a Function with at Most Finite Number of Discontinuities *EXPERIMENTAL*}
\usage{
integrate_discont_val(f, from, to, discontinuities = numeric(0), ...)
}
\arguments{
\item{f}{an R function taking a numeric vector of length 1 as its first
argument and returning a numeric vector of length 1}

\item{from}{the lower limit of integration}

\item{to}{the upper limit of integration}

\item{discontinuities}{nondecreasingly sorted numeric vector which indicates
the points at which \code{f} is discontinuous}

\item{...}{further arguments to be passed to the \code{\link{integrate}} function.}
}
\value{
Returns the estimate of the integral.
}
\description{
The function uses multiple calls to \code{\link{integrate}}.
}
