% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-02-PowerFuzzyNumber.R
\name{PowerFuzzyNumber}
\alias{PowerFuzzyNumber}
\title{Creates a Fuzzy Number with Sides Given by Power Functions}
\usage{
PowerFuzzyNumber(a1, a2, a3, a4, p.left = 1, p.right = 1)
}
\arguments{
\item{a1}{a number specyfing left bound of the support}

\item{a2}{a number specyfing left bound of the core}

\item{a3}{a number specyfing right bound of the core}

\item{a4}{a number specyfing right bound of the support}

\item{p.left}{a positive number specyfing the exponent for the left side}

\item{p.right}{a positive number specyfing the exponent for the right side}
}
\value{
Object of class \code{\linkS4class{PowerFuzzyNumber}}
}
\description{
For convenience, objects of class \code{\linkS4class{PowerFuzzyNumber}}
may be created with this function.
}
\seealso{
Other PowerFuzzyNumber-method: 
\code{\link{Extract}},
\code{\link{PowerFuzzyNumber-class}},
\code{\link{alphaInterval}()},
\code{\link{as.PowerFuzzyNumber}()},
\code{\link{as.TrapezoidalFuzzyNumber}()},
\code{\link{as.character}()},
\code{\link{expectedInterval}()}
}
\concept{PowerFuzzyNumber-method}
