\docType{methods}
\name{show}
\alias{show}
\alias{show,FuzzyNumber-method}
\alias{show,PiecewiseLinearFuzzyNumber-method}
\alias{show,PowerFuzzyNumber-method}
\alias{show,TrapezoidalFuzzyNumber-method}
\title{Print basic information on a fuzzy number}
\description{
  Print basic information on a fuzzy number
}
\section{Methods}{
  \describe{ \item{\code{signature(object =
  "FuzzyNumber")}}{ } \item{\code{signature(object =
  "TrapezoidalFuzzyNumber")}}{ }
  \item{\code{signature(object =
  "PiecewiseLinearFuzzyNumber")}}{ }
  \item{\code{signature(object = "PowerFuzzyNumber")}}{ } }
}
\seealso{
  Other FuzzyNumber.method: \code{\link{alphacut}},
  \code{\link{alphaInterval}}, \code{\link{alphaInterval}},
  \code{\link{alphaInterval}}, \code{\link{alphaInterval}},
  \code{\link{ambiguity}}, \code{\link{Arithmetic}},
  \code{\link{Arithmetic}}, \code{\link{Arithmetic}},
  \code{\link{Arithmetic}}, \code{\link{Arithmetic}},
  \code{\link{Arithmetic}}, \code{\link{convert.side}},
  \code{\link{core}}, \code{\link{distance}},
  \code{\link{distance}}, \code{\link{distance}},
  \code{\link{distance}}, \code{\link{evaluate}},
  \code{\link{expectedInterval}},
  \code{\link{expectedInterval}},
  \code{\link{expectedInterval}},
  \code{\link{expectedInterval}},
  \code{\link{expectedValue}}, \code{\link{Extract}},
  \code{\link{Extract}}, \code{\link{Extract}},
  \code{\link{Extract}}, \code{\link{FuzzyNumber-class}},
  \code{\link{integrateAlpha}},
  \code{\link{integrateAlpha}},
  \code{\link{piecewiseLinearApproximation}},
  \code{\link{plot}}, \code{\link{plot}},
  \code{\link{plot}}, \code{\link{supp}},
  \code{\link{trapezoidalApproximation}},
  \code{\link{value}}, \code{\link{weightedExpectedValue}},
  \code{\link{width}}
}

