\name{TrapezoidalFuzzyNumber}
\alias{TrapezoidalFuzzyNumber}
\title{Creates a trapezoidal fuzzy number}
\usage{
  TrapezoidalFuzzyNumber(a1, a2, a3, a4)
}
\arguments{
  \item{a1}{a number specyfing left bound of the support}

  \item{a2}{a number specyfing left bound of the core}

  \item{a3}{a number specyfing right bound of the core}

  \item{a4}{a number specyfing right bound of the support}
}
\value{
  Object of class
  \code{\linkS4class{TrapezoidalFuzzyNumber}}
}
\description{
  For convenience, objects of class
  \code{\linkS4class{TrapezoidalFuzzyNumber}} may be
  created with this function.
}

