% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/methods-Extract.R
\docType{methods}
\name{Extract}
\alias{Extract}
\alias{[,DiscontinuousFuzzyNumber,character-method}
\alias{[,FuzzyNumber,character-method}
\alias{[,PiecewiseLinearFuzzyNumber,character-method}
\alias{[,PowerFuzzyNumber,character-method}
\title{FuzzyNumber Slot Accessors}
\usage{
\S4method{[}{FuzzyNumber,character}(x, i)

\S4method{[}{PiecewiseLinearFuzzyNumber,character}(x, i)

\S4method{[}{PowerFuzzyNumber,character}(x, i)

\S4method{[}{DiscontinuousFuzzyNumber,character}(x, i)
}
\arguments{
\item{x}{a fuzzy number}

\item{i}{character; slot name}
}
\value{
Returns the slot value.
}
\description{
For possible slot names see man pages for the \linkS4class{FuzzyNumber} class and its derivatives
}
\details{
All slot accessors are read-only.
}
\examples{
A <- FuzzyNumber(1,2,3,4)
A["a1"]
A["right"]
}

