\docType{methods}
\name{show}
\alias{show}
\alias{show,FuzzyNumber-method}
\title{Print Basic Information on a Fuzzy Number}
\usage{
  \S4method{show}{FuzzyNumber}(object)
}
\arguments{
  \item{object}{a fuzzy number}
}
\value{
  nothing interesting
}
\description{
  See \code{\link{as.character}} for more details.
}
\details{
  The method \code{\link{as.character}} is called on given
  fuzzy number object with default arguments. The results
  are printed on \code{stdout}.
}
\seealso{
  Other FuzzyNumber-method: \code{\link{alphacut}},
  \code{\link{alphaInterval}}, \code{\link{ambiguity}},
  \code{\link{Arithmetic}}, \code{\link{as.character}},
  \code{\link{as.FuzzyNumber}},
  \code{\link{as.PiecewiseLinearFuzzyNumber}},
  \code{\link{as.PowerFuzzyNumber}},
  \code{\link{as.TrapezoidalFuzzyNumber}},
  \code{\link{core}}, \code{\link{distance}},
  \code{\link{evaluate}}, \code{\link{expectedInterval}},
  \code{\link{expectedValue}}, \code{\link{FuzzyNumber}},
  \code{\link{FuzzyNumber-class}},
  \code{\link{integrateAlpha}},
  \code{\link{piecewiseLinearApproximation}},
  \code{\link{plot}}, \code{\link{supp}},
  \code{\link{trapezoidalApproximation}},
  \code{\link{value}}, \code{\link{weightedExpectedValue}},
  \code{\link{width}}
}

