\docType{class}
\name{TrapezoidalFuzzyNumber-class}
\alias{TrapezoidalFuzzyNumber-class}
\title{S4 class representing a trapezoidal fuzzy number}
\description{
  S4 class representing a trapezoidal fuzzy number
}
\section{Extends}{
  Class \code{\linkS4class{FuzzyNumber}}, directly.
}
\examples{
showClass("DiscontinuousFuzzyNumber")
}
\seealso{
  \code{\link{TrapezoidalFuzzyNumber}} for a convenient
  constructor
}

