% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-02-TrapezoidalFuzzyNumber.R
\name{TrapezoidalFuzzyNumber}
\alias{TrapezoidalFuzzyNumber}
\title{Creates a Trapezoidal Fuzzy Number}
\usage{
TrapezoidalFuzzyNumber(a1, a2, a3, a4)
}
\arguments{
\item{a1}{a number specyfing left bound of the support}

\item{a2}{a number specyfing left bound of the core}

\item{a3}{a number specyfing right bound of the core}

\item{a4}{a number specyfing right bound of the support}
}
\value{
Object of class \code{\linkS4class{TrapezoidalFuzzyNumber}}
}
\description{
For convenience, objects of class \code{\linkS4class{TrapezoidalFuzzyNumber}}
may be created with this function.
}
\seealso{
Other TrapezoidalFuzzyNumber-method: \code{\link{Arithmetic}},
  \code{\link{TrapezoidalFuzzyNumber-class}},
  \code{\link{TriangularFuzzyNumber}},
  \code{\link{alphaInterval}},
  \code{\link{as.PiecewiseLinearFuzzyNumber}},
  \code{\link{as.PowerFuzzyNumber}},
  \code{\link{as.TrapezoidalFuzzyNumber}},
  \code{\link{expectedInterval}}, \code{\link{plot}}
}
\concept{TrapezoidalFuzzyNumber-method}
