% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-02-TrapezoidalFuzzyNumber.R
\name{TriangularFuzzyNumber}
\alias{TriangularFuzzyNumber}
\title{Creates a Triangular Fuzzy Number}
\usage{
TriangularFuzzyNumber(a1, amid, a4)
}
\arguments{
\item{a1}{a number specyfing left bound of the support}

\item{amid}{a number specyfing the core}

\item{a4}{a number specyfing right bound of the support}
}
\value{
Object of class \code{\linkS4class{TrapezoidalFuzzyNumber}}
}
\description{
For convenience, objects of class \code{\linkS4class{TrapezoidalFuzzyNumber}}
may be created with this function.
}
\details{
Currently there is no separate class of a Triangular Fuzzy Number.
}
\seealso{
Other TrapezoidalFuzzyNumber-method: \code{\link{Arithmetic}},
  \code{\link{TrapezoidalFuzzyNumber-class}},
  \code{\link{TrapezoidalFuzzyNumber}},
  \code{\link{alphaInterval}},
  \code{\link{as.PiecewiseLinearFuzzyNumber}},
  \code{\link{as.PowerFuzzyNumber}},
  \code{\link{as.TrapezoidalFuzzyNumber}},
  \code{\link{expectedInterval}}, \code{\link{plot}}
}
\concept{TrapezoidalFuzzyNumber-method}
